Fix preview hook usage in seller transaction listing
- move the dependent transaction-entry fetch into a dedicated Preview label component
- keep Preview focused on resolving the latest transactionEntryId from the collection query
- preserve the existing seller transaction preview output while removing the conditional hook hazard
